createDeveloper class for crud

+++ src/main/java/com/workintech/sprintday2/model/Developer.java
@@ -5,10 +5,10 @@ import java.util.Objects;
 public class Developer {
     private int id;
     private String name;
-    private int salary;
+    private double salary;
     private Experience experience;
 
-    public Developer(int id, String name, int salary, Experience experience) {
+    public Developer(int id, String name, double salary, Experience experience) {
         this.id = id;
         this.name = name;
         this.salary = salary;
@@ -19,8 +19,8 @@ public class Developer {
     public void setId(int id) {this.id = id;}
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
-    public int getSalary() {return salary;}
-    public void setSalary(int salary) {this.salary = salary;}
+    public double getSalary() {return salary;}
+    public void setSalary(double salary) {this.salary = salary;}
     public Experience getExperience() {return experience;}
     public void setExperience(Experience experience) {this.experience = experience;}
