+++ src/main/java/com/workintech/sprintday2/model/JuniorDev.java
@@ -1,7 +1,7 @@
 package com.workintech.sprintday2.model;
 
 public class JuniorDev extends Developer {
-    public JuniorDev(int id, String name, int salary) {
+    public JuniorDev(int id, String name, double salary) {
 
         super(id, name, salary, Experience.JUNIOR);
     }
